ignorer le dossier image, middleware auth sur route

+++ backend/controllers/sauce.js
@@ -1,6 +1,7 @@
 const Sauce = require ('../models/Sauce');
+const fs = require('fs');
 //*******************************************création sauce**************************************** */
-exports.createSauces = (req,res,next) => {
+exports.createSauce = (req,res,next) => {
     const sauceObject = JSON.parse(req.body.sauce);
     delete sauceObject._id;
     delete sauceObject._userId;
@@ -9,7 +10,7 @@ exports.createSauces = (req,res,next) => {
         userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-
+    console.log(sauceObject)
     sauce.save()
     .then(()=>{ res.status(201).json({message:'Sauce enregistrée'})})
     .catch(error=> { res.status(400).json({error})})
@@ -17,16 +18,42 @@ exports.createSauces = (req,res,next) => {
 
 //*******************************************modification sauce**************************************** */
 exports.modifySauce = (req,res,next)=>{
-    Sauce.updateOne({_id: req.params.id}, {...req.body,_id: req.params.id})
-    .then(()=> res.status(200).json({message: 'Sauce modifié'}))
-    .catch(error => res.status(400).json({error}));
+  const sauceObject = req.file ? {
+    ...JSON.parse(req.body.sauce) ,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : {...req.body};
+
+  delete sauceObject._userId;
+  Sauce.findOne({_id: req.params.id})
+  .then((sauce)=>{
+    if (sauce.userId != req.auth.userId){
+        res.status(401).json({message:'non-autorisé'});
+    } else {
+        Sauce.updateOne({_id:req.params.id},{...sauceObject, _id: req.params.id})
+        .then(()=> res.status(200).json({message: 'Sauce modifié'}))
+        .catch(error=> res.status(401).json({error}));
+    }
+  })
+  .catch((error)=>{
+    res.status(400).json({error})
+  })
 };
 //*******************************************suppréssion sauce**************************************** */
 
 exports.deleteSauce = (req,res,next)=>{
-    Sauce.deleteOne({ _id: req.params.id })
-    .then(()=> res.status(200).json({ message: 'Sauce suprimé'}))
-    .catch(error=> res.status(400).json({ error }))
+    Sauce.findOne({_id: req.params.id})
+    .then(sauce => {
+      if (sauce.userId != req.auth.userId){
+        res.status(401).json({message: 'Non autorisé'});
+      } else {
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, ()=>{
+          Sauce.deleteOne({_id:req.params.id})
+          .then(()=>{res.status(200).json({message: 'objet suprimé !'})})
+          .catch(error => res.status(401).json({error}))
+        })
+      }
+    })
 };
 //*******************************************sélectioné une sauce**************************************** */
 exports.getOneSauce = (req,res,next)=>{
@@ -37,7 +64,7 @@ exports.getOneSauce = (req,res,next)=>{
 //*******************************************retourne le tableau de l'ensemble des sauces**************************************** */
 exports.getAllSauces = (req,res,next) => {
     Sauce.find()
-    .then(sauces=>res.status(200).json(sauces))
+    .then(sauce=>res.status(200).json(sauce))
     .catch(error => res.status(400).json({error}))
 }
